# specs/openapi.yaml
openapi: 3.1.0
info:
  title: KINSHASA MOSOLO — API du socle (v1)
  version: 1.0.0
  summary: Plateforme souveraine des recettes de la Ville-Province de Kinshasa
  description: |
    Contrat de référence : `specs/contrat-api.md`. Montants au format `MoneyJSON` (`{"amount": "150.00", "currency": "USD"}`),
    jamais de nombre flottant. Erreurs au format RFC 9457 (`application/problem+json`) avec un `code` stable.

    **Authentification de démonstration** : en-tête `x-demo-user: <userId>` (voir `GET /v1/demo/users`).
    En production : OIDC + clés d'accès (passkeys), mTLS + signatures pour les prestataires.

    Routes marquées `x-extension: true` : compléments du socle, non listés dans le contrat v1 (le frontend peut les ignorer).
  license:
    name: Propriété de la Ville-Province de Kinshasa
servers:
  - url: http://localhost:8080
    description: Développement local (`npm run dev -w backend`)
...
security:
  - demoUser: []
paths:
...
  /v1/settlements/statements:
    post:
      tags: [Trésor]
      summary: Import d'un relevé (R17) → rapprochement à trois voies
      description: Idempotent par `statementId` (rejeu identique → 200 ; contenu différent → 409 `STATEMENT_ALREADY_IMPORTED`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatementRequest' }
      responses:
        '201':
          description: Relevé importé et rapproché
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatementResult' }
        '200':
          description: Relevé déjà importé (rejeu)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatementResult' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /v1/reconciliation/exceptions:
    get:
      tags: [Trésor]
      summary: Exceptions de rapprochement (R17, R18, R22)
      responses:
        '200':
          description: Exceptions ouvertes (dont « règlement manquant » calculé à J+1)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReconciliationException' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/ledger/entries:
    get:
      tags: [Trésor]
      summary: Écritures du grand livre (R17, R22, R23)
      parameters:
        - { name: sourceId, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Écritures (ajout seul, chaînées)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LedgerEntry' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/ledger/entries/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    delete:
      tags: [Trésor]
      summary: Interdit — grand livre en ajout seul
      x-extension: true
      responses:
        '405': { $ref: '#/components/responses/Problem' }
    put:
      tags: [Trésor]
      summary: Interdit — grand livre en ajout seul
      x-extension: true
      responses:
        '405': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Trésor]
      summary: Interdit — grand livre en ajout seul
      x-extension: true
      responses:
        '405': { $ref: '#/components/responses/Problem' }
  /v1/ledger/entries/{id}/reversals:
    post:
      tags: [Trésor]
      summary: Contre-écriture liée à l'original (R17) — seule forme de correction
      x-extension: true
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              additionalProperties: false
              properties:
                reason: { type: string, minLength: 5 }
      responses:
        '201':
          description: Contre-écriture passée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LedgerEntry' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /v1/ledger/balance:
    get:
      tags: [Trésor]
      summary: Équilibre du grand livre (R17, R22, R23)
      responses:
        '200':
          description: Équilibre par devise et soldes par compte
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LedgerBalance' }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    demoUser:
      type: apiKey
      in: header
      name: x-demo-user
      description: Identifiant d'un utilisateur de démonstration (voir GET /v1/demo/users). DÉMO UNIQUEMENT.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Problem:
      description: Erreur RFC 9457
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          description: Code stable
          examples: [RULE_NOT_EXECUTABLE, FORBIDDEN, SEPARATION_OF_DUTIES, IDEMPOTENCY_KEY_REUSED, NONCE_REPLAYED, FX_RATE_MISSING, PROVIDER_UNAVAILABLE, AMOUNT_NOT_REPRESENTABLE, PROVIDER_ORDER_MISMATCH]
        instance: { type: string }
      additionalProperties: true
    CurrencyCode:
      type: string
      enum: [CDF, USD, EUR, GBP, CAD, CHF, ZAR, XAF, AOA, ZMW, RWF, UGX, KES, TZS, BIF, CNY, AED]
...
    MoneyJSON:
      type: object
      required: [amount, currency]
      additionalProperties: false
      properties:
        amount: { type: string, pattern: '^\d+(\.\d+)?$', examples: ['150.00'] }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
...
    StatementRequest:
      type: object
      required: [statementId, lines]
      additionalProperties: false
      properties:
        statementId: { type: string }
        lines:
          type: array
          minItems: 1
          items:
            type: object
            required: [accountAlias, amount, valueDate, paymentReference]
            additionalProperties: false
            properties:
              accountAlias: { type: string }
              amount: { $ref: '#/components/schemas/MoneyJSON' }
              valueDate: { type: string, format: date }
              paymentReference: { type: string }
    StatementResult:
      type: object
      properties:
        statementId: { type: string }
        importedAt: { type: string, format: date-time }
        importedBy: { type: string }
        lines: { type: integer }
        matched:
          type: array
          items:
            type: object
            properties:
              paymentReference: { type: string }
              receiptNumber: { type: string }
              obligationId: { type: string }
              amount: { $ref: '#/components/schemas/MoneyJSON' }
        exceptions: { type: array, items: { $ref: '#/components/schemas/ReconciliationException' } }
    ReconciliationException:
      type: object
      required: [id, type, detail, status, openedAt]
      properties:
        id: { type: string }
        type: { type: string, enum: [ORPHAN_CREDIT, CREDIT_WITHOUT_CONFIRMATION, DUPLICATE_CREDIT, WRONG_ACCOUNT, UNKNOWN_ACCOUNT, AMOUNT_MISMATCH, MISSING_SETTLEMENT, PROVIDER_AMBIGUOUS] }
        statementId: { type: string }
        paymentReference: { type: string }
        line: { type: object, additionalProperties: true }
        detail: { type: string }
        status: { type: string, const: OUVERTE }
        openedAt: { type: string, format: date-time }
        computed: { type: boolean }
    LedgerEntry:
      type: object
      required: [id, seq, at, eventType, lines, prevHash, hash]
      properties:
        id: { type: string }
        seq: { type: integer }
        at: { type: string, format: date-time }
        eventType: { type: string, examples: [ASSESSMENT, PAYMENT_CONFIRMED, SETTLEMENT_CREDITED, REVERSAL] }
        description: { type: string }
        sourceType: { type: string }
        sourceId: { type: string }
        lines:
          type: array
          items:
            type: object
            properties:
              account: { type: string, enum: [CREANCES_CONTRIBUABLES, RECETTES_CONSTATEES, FONDS_A_RECEVOIR_PRESTATAIRES, COMPTE_PUBLIC_RECETTES] }
              side: { type: string, enum: [DEBIT, CREDIT] }
              amount: { $ref: '#/components/schemas/MoneyJSON' }
        reversalOf: { type: string }
        reason: { type: string }
        prevHash: { type: string }
        hash: { type: string }
    LedgerBalance:
      type: object
      required: [balanced, entries, byCurrency, accounts]
      properties:
        balanced: { type: boolean }
        entries: { type: integer }
        byCurrency:
          type: array
          items:
            type: object
            properties:
              currency: { $ref: '#/components/schemas/CurrencyCode' }
              debit: { $ref: '#/components/schemas/MoneyJSON' }
              credit: { $ref: '#/components/schemas/MoneyJSON' }
              balanced: { type: boolean }
        accounts:
          type: array
          items:
            type: object
            properties:
              account: { type: string }
              label: { type: string }
              currency: { $ref: '#/components/schemas/CurrencyCode' }
              balance: { $ref: '#/components/schemas/MoneyJSON' }
        headHash: { type: [string, 'null'] }
...